20240114_IT#0811_Azure_OCR_使い方/requirements.txt
azure-cognitiveservices-vision-computervision
msrest